refactor: programs renamed to ensure clarity, some extraneous code removed
application was also tested with some small bugs removed

+++ riptide_database/src/schema.rs
@@ -1,6 +1,6 @@
 table! {
     shares (file_id) {
-        file_id -> Integer,
+        file_id -> BigInt,
         exp -> BigInt,
         crt -> BigInt,
         file_size -> BigInt,
